+++ src/render/vulkan/VKCall.h
@@ -31,6 +31,7 @@ public:
         ScriptEngine &scriptEngine);
 
 private:
+    bool validateShaderTypes();
     int evaluateInt(ScriptEngine &scriptEngine, const QString &expression);
     uint32_t evaluateUInt(ScriptEngine &scriptEngine,
         const QString &expression);