+++ src/session/Item.h
@@ -308,6 +308,8 @@ int getFieldRowOffset(const Field &field);
 int getBlockStride(const Block &block);
 TextureKind getKind(const Texture &texture);
 CallKind getKind(const Call &call);
+bool callTypeSupportsShaderType(Call::CallType callType,
+    Shader::ShaderType shaderType);
 bool shouldExecute(Call::ExecuteOn executeOn, EvaluationType evaluationType);
 
 SourceType getSourceType(Shader::ShaderType type, Shader::Language language);