+++ CHANGELOG.md
@@ -17,6 +17,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Moved `selection` from app to session script object.
 - Keyboard script object `keys` array can be indexed by ASCII code.
+- Synchronize shader type in session with when saving editor as.
+- Setting default call item name to call type.
+
+### Fixed
+
+- Validating shader types for calls.
 
 ## [Version 2.5.0] - 2025-04-25